backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Vector Database
qdrant-client==1.7.3

# Cache
redis[hiredis]==5.0.1

# AI/ML
openai==1.12.0
google-generativeai==0.3.2
tiktoken==0.6.0
fastembed==0.2.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.2
beautifulsoup4==4.12.3
python-magic==0.4.27

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# OAuth & Auth
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.117.0
PyGithub==2.2.0

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0
faker==23.1.0

# Development
black==24.2.0
ruff==0.2.1
mypy==1.8.0
pre-commit==3.6.1

# Load Testing
locust==2.20.0
